Switched to constant salt value for Scrambling
Removed reservation on 11 bytes for salt storage, placing at beginning
of image makes it too easy to destroy the salt. Going w/ a static salt
that was generated separately by Rfc2898DeriveBytes.

The bytes used in this instance are not for cryptographic purposes but
rather for random placing for the data pixels.